+++ lib/equire-prof.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require "require_profiler"
+
+RequireProfiler.start
+at_exit { RequireProfiler.stop }